Split up some big files.
